pyproject.toml
[tool.poetry]
name = "daily-scrum-bot"
version = "0.0.1"
description = "A Telegram Bot to manage Daily Scrum"
authors = ["educup <hello@educup.io>"]
homepage = "https://github.com/educup/daily-scrum-bot"
repository = "https://github.com/educup/daily-scrum-bot"
documentation = "https://github.com/educup/daily-scrum-bot"
readme = "README.md"
license = "MIT"

[tool.poetry.scripts]
ds-bot = "daily_scrum_bot.main:app"

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.4.0"
beanie = "^1.8.8"
python-dotenv = "^0.19.2"
dnspython = "^2.1.0"
python-telegram-bot = "^13.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-asyncio = "^0.15.1"
flake8 = "^3.9.2"
black = "^21.7b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
